Adding more tree fuctionality
