+++ requirements.txt
@@ -1,3 +1,4 @@
+scikit-learn
 pandas
 tree
 torch
